Improve speed efficiency
Apparenty the @specializations don't avoid most of the boxing, so I decided to avoid them manually replacint all the functional implementations of the parsers with an imperative one.
The resulting parsing code is four times faster than the previos commit, at the cost of elegance.
Anter there is still room for more improvements.