Make `Appender.append` and `Parser.parse` never throw an exception.
Previously the `Parser.parse` for the `Nothing` type was throwing an exception. Now it calls `Cursor.faile` instead.

Previouly the `Appender.append` for coproducts was throwing an exception when the parameter actual type is not contained in the `CoproductAppender.productsInfo` array. Now it calls `Record.fail` instead.

Add the `fail` and `failures` methods to `Record`.

Add the `ToJsonConvertible.toJsDocument` method.

Make JSON abstract syntax tree types extend AnyVal.